refactor(benchmarks): localize workload scene behavior
