+++ apps/benchmarks/src/workloads/dynamic-layout.ts
@@ -4,7 +4,11 @@ import * as THREE from 'three/webgpu';
 import { benchmarkContentWidth, LIVE_TEXT_COLOR, LIVE_TEXT_LINE_HEIGHT } from './shared/text-style';
 
 import type { ComparisonWorkloadConfiguration, ComparisonWorkloadDefinition } from './contracts';
-import type { ComparisonWorkloadEntry, WorkloadTextFactoryContext } from './factory-contracts';
+import {
+  committedTextLayout,
+  type ComparisonWorkloadEntry,
+  type WorkloadTextFactoryContext,
+} from './factory-contracts';
 
 export const DYNAMIC_LAYOUT_TEXT = [
   'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Left-aligned lines narrow and open while every word reshapes into its changing measure.',
@@ -85,6 +89,64 @@ export function setDynamicLayoutBoundsVisibility(bounds: { visible: boolean }, v
   bounds.visible = visible;
 }
 
+export function animateDynamicLayoutEntries(
+  entries: readonly ComparisonWorkloadEntry[],
+  configuration: Pick<
+    ComparisonWorkloadConfiguration,
+    'amount' | 'animationEnabled' | 'animationSpeed' | 'layoutWidthRatio'
+  >,
+  timestamp: number,
+  viewportWidth: number,
+  viewportHeight: number,
+  widthsScratch: Float64Array,
+  onError: (error: unknown) => void,
+  onReflow: (duration: number) => void,
+): void {
+  if (!configuration.animationEnabled) return;
+  const nextWidths = dynamicLayoutWidths(configuration, viewportWidth, timestamp, widthsScratch);
+  if (
+    entries.length === nextWidths.length &&
+    entries.every((entry, index) => entry.lastWidth !== undefined && Math.abs(nextWidths[index]! - entry.lastWidth) < 1)
+  ) {
+    return;
+  }
+  const reflowStarted = performance.now();
+  try {
+    for (const [index, entry] of entries.entries()) {
+      entry.lastWidth = nextWidths[index]!;
+      if (entry.widthUpdate === undefined) throw new Error('dynamic layout entry is missing its retained width update');
+      entry.widthUpdate.width = nextWidths[index]!;
+      entry.text.setProperties(entry.widthUpdate);
+    }
+    for (const { node } of entries) node.updateMatrixWorld(true);
+    layoutDynamicLayoutEntries(entries, viewportWidth, viewportHeight);
+    onReflow(performance.now() - reflowStarted);
+  } catch (error) {
+    onError(error);
+  }
+}
+
+export function layoutDynamicLayoutEntries(
+  entries: readonly ComparisonWorkloadEntry[],
+  viewportWidth: number,
+  viewportHeight: number,
+): void {
+  const inset = 20;
+  const laneHeight = viewportHeight / Math.max(1, entries.length);
+  for (const [index, entry] of entries.entries()) {
+    const layout = committedTextLayout(entry.text);
+    const x =
+      entry.alignment === 'end'
+        ? viewportWidth - inset - layout.width
+        : entry.alignment === 'center'
+          ? (viewportWidth - layout.width) / 2
+          : inset;
+    const y = index * laneHeight + Math.max(inset, (laneHeight - layout.height) / 2);
+    entry.text.position.set(x, -y, 0);
+    if (entry.bounds !== undefined) updateLayoutBounds(entry.bounds, x, y, layout.width, layout.height);
+  }
+}
+
 function createLayoutBounds(): THREE.LineSegments<THREE.BufferGeometry, THREE.LineBasicNodeMaterial> {
   const geometry = new THREE.BufferGeometry();
   geometry.setAttribute('position', new THREE.BufferAttribute(new Float32Array(8 * 3), 3));
@@ -96,6 +158,51 @@ function createLayoutBounds(): THREE.LineSegments<THREE.BufferGeometry, THREE.Li
     transparent: true,
   });
   return new THREE.LineSegments(geometry, material);
+}
+
+function updateLayoutBounds(
+  bounds: THREE.LineSegments<THREE.BufferGeometry, THREE.LineBasicNodeMaterial>,
+  x: number,
+  y: number,
+  width: number,
+  height: number,
+): void {
+  const positions = bounds.geometry.getAttribute('position');
+  const right = x + width;
+  const bottom = -(y + height);
+  const top = -y;
+  const vertices = [
+    x,
+    top,
+    0,
+    right,
+    top,
+    0,
+    right,
+    top,
+    0,
+    right,
+    bottom,
+    0,
+    right,
+    bottom,
+    0,
+    x,
+    bottom,
+    0,
+    x,
+    bottom,
+    0,
+    x,
+    top,
+    0,
+  ];
+  if (!(positions.array instanceof Float32Array)) {
+    throw new TypeError('dynamic layout bounds require a Float32 position buffer');
+  }
+  positions.array.set(vertices);
+  positions.needsUpdate = true;
+  bounds.geometry.computeBoundingSphere();
 }
 
 function animationRate(animationSpeed: number): number {
